Add files via upload
The Netflix EDA project provided insights into the movies and TV shows on Netflix. It was observed that Drama, Comedy, and Action are the top genres on the platform, with the United States producing the most content. The number of movies on Netflix has increased significantly in recent years, with the 2010s being the decade with the most movie releases. TV shows, on the other hand, have been more consistently released over time. TV-MA is the most popular rating on Netflix, and the majority of movies on the platform have a duration of 90-120 minutes.

Overall, the project demonstrated the power of exploratory data analysis in providing insights into large datasets, and how visualizations can aid in communicating these insights.